pyproject.toml
[project]
name = "liquid_os_example_agent"
version = "0.1.0"
description = "Liquid OS Example Agent"
authors = [{name = "Yash Agarwal", email = "yash@nurturelabs.co"}]
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.112.1",
    "tortoise-orm",
    "web3",
    "aiohttp>=3.9.3",
    "sentry-sdk>=1.39.1",
    "simplejson>=3.19.3",
    "pydantic-settings>=2.7.1",
    "eth-account",
    "asyncpg>=0.30.0",
    "orjson>=3.10.15",
    "slowapi>=0.1.9",
    "rich>=13.7.1",
    "hyperliquid-python-sdk==0.9.0",
    "redis>=5.2.1",
    "setuptools>=75.8.0",
    "fastapi-cache2[redis]>=0.1.8",
    "logfire>=3.4.0",
    "logfire-api>=3.4.0",
    "griffe>=1.5.6",
    "openai>=1.61.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=4.1.0",
    "ruff>=0.9.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
